pyproject.toml
[project]
name = "Metamorf"
version = "0.5.0"
authors = [
  { name="Guillermo Aumatell", email="gaumatellsalom@gmail.com" },
]
description = "Metadata Suite to build and manage your Datawarehouse"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "snowflake-connector-python",
    "PyYAML",
    "Flask",
    "psycopg2-binary",
    "mysql-connector-python"
]

[tool.setuptools]
packages = ["metamorf", "metamorf.engines", "metamorf.tools"]

[project.urls]
"Homepage" = "https://github.com/gasalom/metamorf"
"Bug Tracker" = "https://github.com/gasalom/metamorf/issues"

[project.scripts]
metamorf = "metamorf.main:main"

[build-system]
requires = ["setuptools>=61.0", "snowflake-connector-python", "PyYAML", "Flask", "psycopg2-binary"]
build-backend = "setuptools.build_meta"